+++ apps/api/src/app/portfolio/calculator/twr/portfolio-calculator.ts
@@ -19,7 +19,7 @@ export class TwrPortfolioCalculator extends PortfolioCalculator {
     end: Date;
     exchangeRates: { [dateString: string]: number };
     marketSymbolMap: {
-      [date: string]: { [symbol: string]: Big };
+      [date: string]: { [assetProfileIdentifier: string]: Big };
     };
     start: Date;
     step?: number;